site_proxy: support regex in wait_for_status_update
Make wait_for_status_update more flexible by supporting regex.
Until now, the expected status must equal. This works fine as long as
the status is predictable (e.g. "True", "False").

In order to support some statuses in the TLC SXL, such as S0001 or
S0002 which are text strings, where only a single char is relevant for a
given command it is useful to match against a regex instead